Rename `pool` to `blox`
rename `pool` to `blox` to tighten up terminology.
